requirements.txt
pyyaml
python-dotenv
pydantic
requests